refactored timescale package
