+++ .gitignore
@@ -4,3 +4,4 @@ build/
 django_timescaledb.egg-info/
 .env
 *ipynb
+example/timescale
