--- timescale/db/models/__init__.py
@@ -1,7 +1,0 @@
-from timescale.db.models.models import __all__ as all_models
-from timescale.db.models.querysets import __all__ as all_querysets
-from timescale.db.models.managers import __all__ as all_managers
-from timescale.db.models.fields import __all__ as all_fields
-
-
-__all__ = all_models + all_querysets + all_managers + all_fields
