+++ timescale/db/models/expressions.py
@@ -1,5 +1,5 @@
 from django.db import models
-from django.db.models.aggregates import Aggregate
+from django.contrib.postgres.fields import ArrayField
 from django.db.models.functions.mixins import (
     FixDurationInputMixin, NumericOutputFieldMixin,
 )
@@ -7,6 +7,7 @@ from django.db.models.functions.mixins import (
 
 class TimeBucket(models.Func):
     function = 'time_bucket'
+    name = "time_bucket"
 
     def __init__(self, expression, interval):
         if not isinstance(interval, models.Value):
@@ -14,11 +15,11 @@ class TimeBucket(models.Func):
         super().__init__(interval, expression)
 
 
-class Bucket(NumericOutputFieldMixin, Aggregate):
-    function = 'time_bucket'
-    allow_distinct = True
+class Histogram(models.Aggregate):
+    function = 'histogram'
+    name = 'histogram'
+    output_field = ArrayField(models.FloatField())
 
-    def __init__(self, expression, interval):
-        if not isinstance(interval, models.Value):
-            interval = models.Value(interval)
-        super().__init__(interval, expression)
+    def __init__(self, expression, min_value, max_value, bucket):
+        super().__init__(expression, min_value, max_value, bucket)
+
